DataBase Done + Fixed Models
Am reparat modelele, relatiile si loop urile dinte ER.

+++ backend-MT/Models/Plata.cs
@@ -8,7 +8,7 @@ namespace backend_MT.Models
     {
         [Key]
         public int plataId { get; set; }
-        public decimal suma { get; set; }
+        public int suma { get; set; }
         public DateTime data { get; set; }
         public int userId { get; set; }
         public virtual User user { get; set; }
